Separate [selectors] ui/selectors
