+++ src/components/AssignmentSchema/Form/MainInfo/MainInfo.tsx
@@ -2,7 +2,7 @@ import { useLocale } from '@hooks/useLocale';
 import { FC, memo, useMemo } from 'react';
 import styles from './mainInfo.module.css';
 import { CustomEditor, TextInput, NumberInput } from '@ui/basics';
-import TagSelector from '@ui/TagSelector/TagSelector';
+import { TagSelector } from '@ui/selectors';
 import stepperStyles from '@styles/ui/stepper.module.css';
 
 const MainInfo: FC<{ form: any }> = ({ form }) => {
