+++ src/components/Group/Form/form.module.css
@@ -1,38 +1,6 @@
-.itemWrapper {
-  padding: var(--spacer-xs) 0;
-  display: flex;
-  flex-direction: row;
-  cursor: pointer;
-}
-
-.actions {
-  margin-left: auto;
-  margin-right: var(--spacer-xs);
-}
-
 .buttonWrapper {
   width: 100%;
   display: flex;
   justify-content: center;
   margin-top: var(--spacer-l);
-}
-
-.nameWrapper {
-  width: 100%;
-  display: flex;
-  justify-content: space-between;
-}
-
-.name {
-  font-size: var(--font-size-m);
-}
-
-.groups {
-  margin-right: var(--spacer-xs);
-  display: flex;
-  flex-direction: row;
-  justify-content: flex-end;
-  align-items: flex-end;
-  font-size: var(--font-size-xs);
-  color: var(--dark3);
 }
