+++ src/components/Tournament/Form/TypeInfo/TypeInfo.tsx
@@ -3,7 +3,7 @@ import { MultiSelect } from '@mantine/core';
 import { Select } from '@ui/basics';
 
 import { FC, memo, useState, useEffect } from 'react';
-import TagSelector from '@ui/TagSelector/TagSelector';
+import { TagSelector } from '@ui/selectors';
 import styles from './typeInfo.module.css';
 import ProgramLanguageSelector from '@ui/ProgramLanguageSelector/ProgramLanguageSelector';
 import { IAssessmentType } from '@custom-types/data/ITournament';
