+++ src/components/ui/selectors/UserSelector/userSelector.module.css
@@ -1,0 +1,31 @@
+.itemWrapper {
+  padding: var(--spacer-xs) 0;
+  justify-content: space-between;
+  display: flex;
+  flex-direction: row;
+  cursor: pointer;
+  width: 100%;
+}
+.nameWrapper {
+  display: flex;
+  justify-content: space-between;
+  width: 100%;
+}
+
+.name {
+  font-size: var(--font-size-m);
+}
+.groups {
+  margin-right: var(--spacer-xs);
+  display: flex;
+  flex-direction: row;
+  justify-content: flex-end;
+  align-items: flex-end;
+  font-size: var(--font-size-xs);
+  color: var(--dark3);
+  max-width: 60%;
+}
+.actions {
+  width: fit-content;
+  margin-right: var(--spacer-xs);
+}
