+++ src/components/ui/selectors/index.ts
@@ -1,0 +1,13 @@
+import TaskSelector from './TaskSelector/TaskSelector';
+import UserSelector from './UserSelector/UserSelector';
+import TagSelector from './TagSelector/TagSelector';
+import GroupSelector from './GroupSelector/GroupSelector';
+import RoleSelector from './RoleSelector/RoleSelector';
+
+export {
+  TaskSelector,
+  UserSelector,
+  TagSelector,
+  GroupSelector,
+  RoleSelector,
+};
