+++ src/locale/en/group/index.ts
@@ -3,7 +3,5 @@ import { form } from './form';
 export const group = {
   add: 'Add group',
   name: 'Title',
-  users: 'Users',
-  selectedUsers: 'Selected users',
   form,
 };
