+++ src/locale/en/ui/userSelector.ts
@@ -1,0 +1,4 @@
+export const userSelector = {
+  selected: 'Selected users',
+  unselected: 'Users',
+};
