+++ src/locale/ru/group/index.ts
@@ -3,7 +3,5 @@ import { form } from './form';
 export const group = {
   add: 'Добавить группу',
   name: 'Название',
-  users: 'Пользователи',
-  selectedUsers: 'Выбранные пользователи',
   form,
 };
