+++ src/locale/ru/ui/index.ts
@@ -1,11 +1,13 @@
-import { codeArea } from "./codeArea"
-import { table } from "./table"
-import { tagSelector } from "./tagSelector"
-import { langSelector } from "./langSelector"
+import { codeArea } from './codeArea';
+import { table } from './table';
+import { tagSelector } from './tagSelector';
+import { langSelector } from './langSelector';
+import { userSelector } from './userSelector';
 
 export const ui = {
-    codeArea,
-    table,
-    tagSelector,
-    langSelector
-}
+  codeArea,
+  table,
+  tagSelector,
+  langSelector,
+  userSelector,
+};
