+++ src/locale/ru/ui/userSelector.ts
@@ -1,0 +1,4 @@
+export const userSelector = {
+  unselected: 'Выбранные пользователи',
+  selected: 'Пользователи',
+};
